\docType{package}
\name{bmetatools-package}
\alias{bmetatools-package}
\title{The 'bmetatools' package.}
\description{
A R package for Bayesian analysis tools for meta-analysis.
}
\references{
Hamaguchi, Y., Matsushima, Y. and Noma, H. (2020). 
Bayesian outlier detection and influence diagnostics in meta-analysis. 
In Preparation.
}
